pyproject.toml
[project]
name = "langgraph_1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gradio>=6.12.0",
    "ipykernel>=7.2.0",
    "ipython>=9.12.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
